Fix bug in login functionality

+++ EstudosPython/CursoEmVideo/Exercícios/03_desafio.py
@@ -2,8 +2,8 @@
 # Crie um script Python que leia dois números e tente mostrar a soma entre eles.
 
 # Tarefa 1: Ler dois números
-primeiro_numero = int(input('Digite o primeiro valor: '))
-segundo_numero = int(input('Digite o segundo valor: '))
+primeiro_numero = float(input('Digite o primeiro valor: '))
+segundo_numero = float(input('Digite o segundo valor: '))
 soma = primeiro_numero + segundo_numero
 
 # Tarefa 2: Mostrar a soma entre eles
